refactored AddWorkoutDialog Composable
